+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/memory/Address.java
@@ -70,7 +70,7 @@ public class Address extends IConst implements ExprInterface {
     }
 
     @Override
-    public int getIntValue(Event e, Model model, Context ctx){
+    public long getIntValue(Event e, Model model, Context ctx){
         return Integer.parseInt(model.getConstInterp(toZ3Int(ctx)).toString());
     }    
 }
